+++ SQL/CREATE TABLE DividendDetails.sql
@@ -1,7 +1,7 @@
 CREATE TABLE DividendDetails (
 	Id int PRIMARY KEY identity(1,1),
 	Ticker VARCHAR(50) null,
-	PayoutPerShare DECIMAL null,
+	PayoutPerShare float null,
 	DividendType VARCHAR(50) null,
 	PayoutFrequency int null,
 	DividendDeclaration date null,
